Off-Platform Project: Abruptly Goblins
In this off-platform project you will be the proprietor of a comic and games store, the Sorcery Society, and you decide to host a game night to play the hot new tabletop RPG Abruptly Goblins! You’ll be writing your own Python functions and using dictionaries to provide a new service to gaming attendees of your shop.

Your beloved clientele want to organize a game night at your comics store! Create a system for organizing and tracking gamer’s availability. Programmatically figure out the best day to host a game night and send out emails to your attendees to let them know when to come. Follow the steps below to get started with your project!