Page Visits Funnel
Page Visits Funnel

Cool T-Shirts Inc. has asked you to analyze data on visits to their website. Your job is to build a funnel, which is a description of how many people continue to the next step of a multi-step process.

In this case, our funnel is going to describe the following process:

* A user visits CoolTShirts.com
* A user adds a t-shirt to their cart
* A user clicks “checkout”
* A user actually purchases a t-shirt